requirements.txt
absl-py==2.0.0
accelerate==0.25.0
addict==2.4.0
aenum==3.1.15
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
attrs==23.1.0
Authlib==1.2.1
basicsr==1.4.2
blendmodes==2023
boto3==1.28.83
botocore==1.31.83
build==1.0.3
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
clean-fid==0.1.35
click==8.1.7
contourpy==1.2.0
cryptography==41.0.3
cycler==0.12.1
dataclass-wizard==0.22.2
deprecation==2.1.0
distlib==0.3.7
dnspython==2.4.1
einops==0.7.0
facexlib==0.3.0
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.12.2
filterpy==1.4.5
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.12.0
ftfy==6.1.3
future==0.18.3
gfpgan==1.3.8
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.24.0
google-auth-oauthlib==1.1.0
gradio==3.41.2
gradio_client==0.5.0
grpcio==1.59.3
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.19.4
idna==3.4
imageio==2.33.0
importlib-metadata==6.9.0
importlib-resources==6.1.1
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonmerge==1.9.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
kornia==0.7.0
lark==1.1.8
lazy_loader==0.3
lightning-utilities==0.10.0
llvmlite==0.41.1
lmdb==1.4.1
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
motor==2.3.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
novita_client==0.4.10
numba==0.58.1
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.23.0
opencv-python==4.8.1.78
orjson==3.9.10
packaging==23.1
pandas==1.3.4
piexif==1.1.3
Pillow==10.1.0
pip-autoremove==0.10.0
pip-check==2.8.1
pip-tools==7.3.0
pipdeptree==2.13.1
pipenv==2023.7.23
platformdirs==3.9.1
protobuf==4.25.1
psutil==5.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
pymongo==3.13.0
pyparsing==3.1.1
pypng==0.20220715.0
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytorch-lightning==2.1.2
pytz==2023.3
PyYAML==6.0.1
qrcode==7.3.1
realesrgan==0.3.0
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
resize-right==0.0.2
rpds-py==0.13.2
rsa==4.9
s3transfer==0.7.0
safetensors==0.4.1
scikit-image==0.22.0
scipy==1.11.4
semantic-version==2.10.0
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.27.0
stripe==7.5.0
sympy==1.12
tb-nightly==2.16.0a20231202
tensorboard-data-server==0.7.2
terminaltables==3.1.10
tf-keras-nightly==2.16.0.dev2023120210
tifffile==2023.9.26
timm==0.9.12
tokenizers==0.13.3
tomesd==0.1.3
tomli==2.0.1
toolz==0.12.0
torch==2.1.1
torchdiffeq==0.2.3
torchmetrics==1.2.1
torchsde==0.2.6
torchvision==0.16.1
tqdm==4.66.1
trampoline==0.1.2
transformers==4.30.2
typing_extensions==4.8.0
urllib3==1.26.16
uvicorn==0.24.0.post1
uvloop==0.17.0
virtualenv==20.24.2
virtualenv-clone==0.5.7
watchfiles==0.19.0
wcwidth==0.2.12
websockets==11.0.3
Werkzeug==3.0.1
yapf==0.40.2
yarl==1.9.3
zipp==3.17.0
